Inject classes which need Activity context without a factory
Creating a factory is not optimal since every time we need a GlideLoader or a Navigator, we would have a new instance every time we call the factory. So in order to have only one instance, we inject them and annotated with @Singleton. As we need the Activity context, we just pass it as a method parameter.